requirements.txt
robotframework==4.1.1
robotframework-pythonlibcore==3.0.0
robotframework-seleniumlibrary==5.1.3
selenium==3.141.0
virtualenv==20.8.0
robotframework-requests==0.9.2
robotframework-jsonlibrary==0.3.1